app/requirements.txt
Flask
mysql-connector
flask-mysqldb


